Implement utils file
