+++ src/utils/utils.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <SFML/Graphics.hpp>
+#include <string>
+
+std::vector<sf::String> wrapText(const sf::String& text, const sf::Font& font, unsigned int characterSize, float boxWidth);
